34-poem-lyrics-generator/requirements.txt
requests>=2.31.0
rich>=13.0.0
click>=8.1.0
PyYAML>=6.0
streamlit>=1.28.0
pytest>=7.0.0
